Adds clean task
Move the list of directories (and files) from the build command to a
configurable variable at the top of the file.

Also adds a clean task the walks the files and directories listed then
unlinks the generated js file if it finds a coffee file.

This level of configuration allows the template to be used in many
different directory and app structures.